correct named object for test, xid with _TEST, abbreviation of names; daivision of module scada-lts-e2e-common into modules: scada-lts-e2e-common-api, scada-lts-e2e-common-impl, scada-lts-e2e-common-core; correct config for verbose logging for chromedriver, added method containsObject(CriteriaObject criteria) and correct acceptAlert default impl in PageContent interface for alert not exists; added assert class for containsObject(CriteriaObject criteria); readable error message; correct current url; reload NavigationPage (session id); reduce runners for test with page; added start test time, and localized legend in email msg; update selenide to 5.17.0, correct containsObject(IdentifierObject identifier) for eventhandler;

+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/page/graphicalviews/DeleteGraphicalViewPageTest.java
@@ -3,20 +3,17 @@ package org.scadalts.e2e.test.impl.tests.page.graphicalviews;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.GraphicalViewCriteria;
 import org.scadalts.e2e.page.impl.criterias.IdentifierObjectFactory;
 import org.scadalts.e2e.page.impl.criterias.identifiers.GraphicalViewIdentifier;
 import org.scadalts.e2e.page.impl.pages.graphicalviews.GraphicalViewsPage;
 import org.scadalts.e2e.test.impl.creators.GraphicalViewObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.StringContains.containsString;
 
-@RunWith(TestWithPageRunner.class)
 public class DeleteGraphicalViewPageTest {
 
     private final GraphicalViewIdentifier viewName = IdentifierObjectFactory.viewName();
@@ -28,14 +25,16 @@ public class DeleteGraphicalViewPageTest {
     @Before
     public void createView() {
         criteria = GraphicalViewCriteria.criteria(viewName);
-        graphicalViewObjectsCreator = new GraphicalViewObjectsCreator(TestWithPageUtil.getNavigationPage(), criteria);
+        graphicalViewObjectsCreator = new GraphicalViewObjectsCreator(TestWithPageUtil.openNavigationPage(), criteria);
         graphicalViewsPageSubject = graphicalViewObjectsCreator.createObjects();
     }
 
     @After
     public void clean() {
-        GraphicalViewsPage page = graphicalViewObjectsCreator.deleteObjects();
-        page.acceptAlertOnPage2();
+        if(graphicalViewObjectsCreator != null) {
+            GraphicalViewsPage page = graphicalViewObjectsCreator.deleteObjects();
+            page.acceptAlertOnPage();
+        }
     }
 
     @Test
