+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/page/graphicalviews/MultiTabGraphicalViewPageTest.java
@@ -4,14 +4,12 @@ import lombok.extern.log4j.Log4j2;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.GraphicalViewCriteria;
 import org.scadalts.e2e.page.impl.criterias.IdentifierObjectFactory;
 import org.scadalts.e2e.page.impl.criterias.identifiers.GraphicalViewIdentifier;
 import org.scadalts.e2e.page.impl.pages.graphicalviews.GraphicalViewsPage;
 import org.scadalts.e2e.page.impl.pages.navigation.NavigationPage;
 import org.scadalts.e2e.test.impl.creators.GraphicalViewObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import java.util.Set;
@@ -19,7 +17,6 @@ import java.util.Set;
 import static org.junit.Assert.assertEquals;
 
 @Log4j2
-@RunWith(TestWithPageRunner.class)
 public class MultiTabGraphicalViewPageTest {
 
     private final GraphicalViewIdentifier viewName = IdentifierObjectFactory.viewName();
@@ -30,15 +27,17 @@ public class MultiTabGraphicalViewPageTest {
     public void createView() {
         logger.info("viewName: {}", viewName.getValue());
         GraphicalViewCriteria criteria = GraphicalViewCriteria.criteria(viewName);
-        graphicalViewObjectsCreator = new GraphicalViewObjectsCreator(TestWithPageUtil.getNavigationPage(), criteria);
+        graphicalViewObjectsCreator = new GraphicalViewObjectsCreator(TestWithPageUtil.openNavigationPage(), criteria);
         graphicalViewsPageSubject = graphicalViewObjectsCreator.createObjects();
     }
 
     @After
     public void clean() {
         NavigationPage.closeAllButOnePage();
-        GraphicalViewsPage page = graphicalViewObjectsCreator.deleteObjects();
-        page.acceptAlertOnPage2();
+        if(graphicalViewObjectsCreator != null) {
+            GraphicalViewsPage page = graphicalViewObjectsCreator.deleteObjects();
+            page.acceptAlertOnPage();
+        }
     }
 
     @Test
