+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/service/storungs/live/GetLivesPaginationAllDataLoggingServiceTest.java
@@ -14,7 +14,6 @@ import org.scadalts.e2e.page.impl.pages.navigation.NavigationPage;
 import org.scadalts.e2e.service.impl.services.storungs.StorungAlarmResponse;
 import org.scadalts.e2e.service.impl.services.storungs.PaginationParams;
 import org.scadalts.e2e.test.impl.creators.StorungsAndAlarmsObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestParameterizedWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import java.text.MessageFormat;
@@ -27,7 +26,7 @@ import static org.junit.Assert.assertTrue;
 import static org.scadalts.e2e.test.impl.utils.StorungsAndAlarmsUtil.*;
 
 @Log4j2
-@RunWith(TestParameterizedWithPageRunner.class)
+@RunWith(Parameterized.class)
 public class GetLivesPaginationAllDataLoggingServiceTest {
 
     @Parameterized.Parameters(name = "{index}: offset: {0}, limit: {1}")
@@ -75,7 +74,7 @@ public class GetLivesPaginationAllDataLoggingServiceTest {
                 .map(a -> DataPointCriteria.noChangeAllDataLogging(a, "0"))
                 .toArray(a -> new DataPointCriteria[10]);
 
-        NavigationPage navigationPage = TestWithPageUtil.getNavigationPage();
+        NavigationPage navigationPage = TestWithPageUtil.openNavigationPage();
 
         storungsAndAlarmsObjectsCreator = new StorungsAndAlarmsObjectsCreator(navigationPage, dataSourceCriteria, points);
         storungsAndAlarmsObjectsCreator.createObjects();
@@ -100,7 +99,8 @@ public class GetLivesPaginationAllDataLoggingServiceTest {
 
     @AfterClass
     public static void clean() {
-        storungsAndAlarmsObjectsCreator.deleteObjects();
+        if(storungsAndAlarmsObjectsCreator != null)
+            storungsAndAlarmsObjectsCreator.deleteObjects();
     }
 
     @Test
